requirements.txt
pandas
streamlit
matplotlib
hypothesis
python-dotenv
google-generativeai
openai
